finish refactor for media-card block. relates to #504.
