+++ src/blocks/media-card/icons.js
