+++ src/blocks/media-card/inspector.js
@@ -2,8 +2,8 @@
  * Internal dependencies
  */
 import applyWithColors from './colors';
-import { BackgroundPanel } from '../../../components/background';
-import DimensionsControl from '../../../components/dimensions-control/';
+import { BackgroundPanel } from '../../components/background';
+import DimensionsControl from '../../components/dimensions-control/';
 
 /**
  * WordPress dependencies
